Separated User Struct into two to utilize ParseSwift
