+++ RizzTracker.xcodeproj/xcuserdata/nicholas.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "8120CFD1-2E93-4B58-A15D-26699CAEEEA0"
+   type = "1"
+   version = "2.0">
+</Bucket>
