+++ RizzTracker/AppDelegate.swift
@@ -12,7 +12,6 @@ import ParseSwift
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
-
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) async -> Bool {
         // Override point for customization after application launch.
